requirements.txt
PyYAML==5.4.1
paramiko==2.7.2